refactor(server): move parse/mod.rs inline test modules to siblings (#3888)
Main took #3897, which grew routes/parse/mod.rs, and this branch's layout
query parameter pushed it to 402 lines with no allowlist row. Splitting
rather than recording a new exemption.

The two inline test modules were the outlier in that file: the other ten
test modules beside them are already ratchet-exempt `_tests.rs` siblings
declared with `#[cfg(test)] mod`. These two now match, which takes the
module to 298 lines and leaves real headroom.

The rebase also brought in #3895's batch-recovery tests, whose calls to
StreamingParquetCacheWriter::new and try_cached_replay needed the layout
argument. Call sites only, both passing the default layout, so the replay
behaviour those tests pin is unchanged.
