+++ apps/server/src/routes/parse/cached_replay_batches_tests.rs
@@ -10,6 +10,7 @@
 //! under the 400-line ratchet.
 
 use super::cached_replay::try_cached_replay;
+use crate::services::ParquetLayout;
 use crate::admission::{Admission, AdmissionCfg};
 use crate::config::Config;
 use crate::routes::parse::parquet::ParquetMetadataHeader;
@@ -95,7 +96,7 @@ async fn seed_cache_from_batches(
     batches: &[Vec<MeshData>],
     total_meshes: usize,
 ) {
-    let mut writer = StreamingParquetCacheWriter::new().unwrap();
+    let mut writer = StreamingParquetCacheWriter::new(ParquetLayout::Flat).unwrap();
     for batch in batches {
         writer.append(batch).unwrap();
     }
@@ -116,7 +117,7 @@ async fn seed_cache_from_batches(
 }
 
 async fn replay_sse_payloads(state: &AppState, cache_key: &str) -> Vec<serde_json::Value> {
-    let response = match try_cached_replay(state, cache_key).await {
+    let response = match try_cached_replay(state, cache_key, ParquetLayout::Flat).await {
         Ok(Some(response)) => response,
         other => panic!("expected a cache hit, got {:?}", other.map(|r| r.is_some())),
     };
